OrWord's are now a (required or not) list of words.
Before they were a list of (required or not) words. But adding a ? at the end of a word in the | group meant making the whole group not required. So now the ? has only to be added at the end of the group:
Now "Hello|Hey?" represents a word that can be "Hello", "Hey" or nothing.
Before every one of those meant what stated above: "Hello?|Hey?", "Hello?|Hey", "Hello|Hey?". So this made the file less readable and created problems for the sentence decomposition function. "Hello?|Hey?" would have been divided in four possible sentences, creating duplicates:
"Hello", "" (Hello, but not taken), "Hey", "" (Hey, but not taken).

So ImportanceWord has been removed and a "required" attribute has been added to OrWord.